pyproject.toml
[project]
name = "adk-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-adk>=0.4.2",
    "fastapi>=0.123.10",
    "google-adk>=1.22.1",
    "litellm>=1.81.0",
    "mcp>=1.25.0",
    "python-ulid[pydantic]>=3.1.0",
    "uvicorn>=0.40.0",
]
